+++ packages/ui/src/views/pages/authorize/supported/Index.vue
@@ -14,7 +14,7 @@
     @request="findItems"
   >
     <template #top-left>
-      <h-button color="primary" icon="mdi-plus" label="新建保护资源" @click="toCreate" />
+      <h-button color="primary" icon="mdi-plus" label="新建支持范围" @click="toCreate" />
     </template>
 
     <template #body-cell-actions="props">
@@ -28,24 +28,20 @@
 
 <script setup lang="ts">
 import type {
-  OAuth2ProtectedResourceMetadataEntity,
-  OAuth2ProtectedResourceMetadataConditions,
-  OAuth2ProtectedResourceMetadataProps,
+  OAuth2SupportedScopeEntity,
+  OAuth2SupportedScopeConditions,
+  OAuth2SupportedScopeProps,
 } from '@herodotus/api';
 import type { QTableColumnProps } from '@/composables/declarations';
 
 import { PAGE_NAME, API } from '@/configurations';
 import { useTable, useDateTime } from '@/composables/hooks';
-import { HDeleteButton, HEditButton, HDenseIconButton, HTable } from '@/components';
 
-defineOptions({
-  name: PAGE_NAME.OAUTH2_PROTECTED_RESOURCE_METADATA,
-  components: { HDeleteButton, HEditButton, HDenseIconButton, HTable },
-});
+defineOptions({ name: PAGE_NAME.OAUTH2_SUPPORTED_SCOPE });
 
 const columns: QTableColumnProps = [
-  { name: 'metadataName', field: 'metadataName', align: 'center', label: 'PRM 名称' },
-  { name: 'metadataCode', field: 'metadataCode', align: 'center', label: 'PRM 代码' },
+  { name: 'scopeName', field: 'scopeName', align: 'center', label: '范围名称' },
+  { name: 'scopeCode', field: 'scopeCode', align: 'center', label: '范围代码' },
   { name: 'description', field: 'description', align: 'center', label: '备注' },
   {
     name: 'updateBy',
@@ -65,13 +61,13 @@ const columns: QTableColumnProps = [
   { name: 'actions', field: 'actions', align: 'center', label: '操作' },
 ];
 
-const rowKey: OAuth2ProtectedResourceMetadataProps = 'metadataId';
+const rowKey: OAuth2SupportedScopeProps = 'scopeId';
 
 const selected = ref([]);
 
 const { defaultFormat } = useDateTime();
 const { tableRows, totalPages, pagination, loading, toEdit, toCreate, deleteItemById, findItems } = useTable<
-  OAuth2ProtectedResourceMetadataConditions,
-  OAuth2ProtectedResourceMetadataEntity
->(API.core.oauth2ProtectedResourceMetadata(), PAGE_NAME.OAUTH2_PROTECTED_RESOURCE_METADATA);
+  OAuth2SupportedScopeConditions,
+  OAuth2SupportedScopeEntity
+>(API.core.oauth2SupportedScope(), PAGE_NAME.OAUTH2_SUPPORTED_SCOPE);
 </script>