+++ shared/form-apis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @herodotus-cloud/form-apis
 
+## 5.1.13
+
+### Patch Changes
+
+- [重构] 重命名 OAuth2 资源标识和 OAuth2 保护资源元数据相关类命名，明晰代码用途减少理解困难
+
 ## 5.1.12
 
 ### Patch Changes